==> Code/1_twitter_api.ipynb <==
# 1. Accessing Twitter API
### Elisabeth Earley

#Import required packages
import tweepy
import json
import csv
import pandas as pd
import datetime
import time

print("Last run: " + str(pd.to_datetime('today')))

#authorize my app to begin using the API!
access_token = 'XXX'
access_token_secret = 'XXX'
consumer_key = 'XXX'
consumer_secret = 'XXX'


auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_token, access_token_secret)

API = tweepy.API(auth, wait_on_rate_limit=True, wait_on_rate_limit_notify=True) #waiting on rate limit automatically sleeps
                                                                                #the system (15 mins) when rate limit has been 
                                                                                #reached.

#create dictionary of Colombian elites and their real name + username
elites = {'Alvaro Uribe Velez':'@AlvaroUribeVel', 'Sergio Fajardo':'@sergio_fajardo', \
'Gustavo Petro':'@petrogustavo', 'Ivan Duque Marquez':'@IvanDuque', 'German Vargas Lleras':'@German_Vargas', \
'Juan Manuel Santos':'@juanmansantos', 'Humberto de la Calle':'@DeLaCalleHum'}
names = elites.keys()  #just their real names
handles = elites.values() #just their twitter usernames

#create a profile for each elite that contains all their user information

profiles = {}
for user in elites:
    profiles[user] = API.get_user(elites[user])

# A) Elite "followers"
[Still under construction...each candidate can hve up to 5 million followers. I can only
grab 5000 ID's at a time and can only use those ID's to pull profiles 900 at at time. It would take 13 days to process 1 million followers. Pulling profiles directly can only be done 200 at a time which would take 52 days for 1 million followers.]
### Saving for future use!

follow_id_dict = {}
for user in elites:
    follow_id_dict[user] = API.followers_ids(elites[user], cursor=-1, count=5000)
    

Example of ID list. From the above loop, each candidate has 5000 follower IDs in their list.

print((follow_id_dict['Gustavo Petro'][0][:20]))

# B) Elite "followings" (aka "friends")
1. Gather ID #'s for the accounts that each elite is following 
2. Use ID's to quickly pull user information
3. Save id, name, screen_name, location, created_at, friends_count, followers_count into csv for each elite

friend_id_dict = {}
for user in elites:
    friend_id_dict[user] = API.friends_ids(elites[user])
    
#Create a dictionary for the total list of 'friends' IDs, where each elite has their own list.

#These are the numbers of IDs that we should be pulling. More for some than others.
for name in names:
    print("Total friends for: " + name)
    print(profiles[name].friends_count)
    print('\n')
    

### Create function that downloads specific variables from each 'friend's' profile, writing the data into a csv file
**Expected time: 2.64 hours**

1. Create empty dictionary to store an empty list for every elite.
2. Create an empty csv ('friends.csv') in sublime to avoid the meta-data created by opening a new file in excel.
3. Using empty csv and encoding in utf-8 to avoid the loop breaking due to emojis, create columns for the variables i'm interested in taking from each 'friend'
4. Loops through the ID numbers for each elite's friends list and grabs the 'friend' profile
5. Allow the loop to append 'friend' profiles to the respective empty list within the main dictionary (userlist)
6. Create a dictionary holding each column's information (row_dict)
7. Allow the csv to be written using the row_dict data 

userlist = {}
with open('friends.csv', 'w', encoding='utf-8') as csvfile:
    fieldnames = ['elite', 'id', 'name', 'screen_name', 'location', 'created_at', 'friends_count', 'followers_count']
    writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
    writer.writeheader()
    
    for name in list(elites.keys()):
        userlist[name] = []
        for id_no in friend_id_dict[name]:
            print(name, id_no)
            friend_data = API.get_user(id_no)
            userlist[name].append(friend_data)
            row_dict = {}
            row_dict['elite'] = name
            row_dict['id'] = id_no
            row_dict['name'] = friend_data.name
            row_dict['screen_name'] = friend_data.screen_name
            row_dict['location'] = friend_data.location
            row_dict['created_at'] = friend_data.created_at
            row_dict['friends_count'] = friend_data.friends_count
            row_dict['followers_count'] = friend_data.followers_count
            writer.writerow(row_dict)
        

#quick unicode data clean
#I want to strip all accents from letters FIRST to avoid unicode errors going into R
#locations were being interpreted into R as "PopayÃ¡n, Colombia", etc. 

import unidecode
with open('friends_cleaned.csv', 'w', encoding='utf-8') as csvfile:
    writer = csv.writer(csvfile, delimiter=',', quotechar='"', quoting=csv.QUOTE_MINIMAL)
    with open('friends.csv', 'r', encoding='utf-8') as friends:
        reader = csv.reader(friends, delimiter=',', quotechar='"')   #quotechar is needed so that the commas in the string don't
        for row in reader:                                           #split the location into two. (e.g. "Bogota, Colombia")
            if len(row) > 0:  #only run this function on non-missing data
                location = row[4]
                cleaned_location = unidecode.unidecode(location).strip('"') #first removes accents, then removes " " in the csv
                print(location, cleaned_location) #make sure it's working right by
                new_row = list(row)               #printing old string and new string
                new_row[4] = cleaned_location
                writer.writerow(new_row)
                
            
    


### ...Time to move to data cleaning and visualization in R
(After waiting 3 hours for the data collecting loop to run and write the csv all the way through! I ran mine overnight.)

# C) Download tweets from time-windows
### Saving for future use!

startDate_1 = datetime.datetime(2014, 5, 1, 0, 0, 0) #elections
endDate_1 = datetime.datetime(2014, 8, 1, 0, 0, 0)

startDate_2 = datetime.datetime(2016, 9, 1, 0, 0, 0) #Plebiscite vote on ratification
endDate_2 = datetime.datetime(2016, 12, 31, 0, 0, 0)

startDate_3 = datetime.datetime(2018, 3, 1, 0, 0, 0) #current elections
endDate_3 = datetime.datetime(2018, 7, 1, 0, 0, 0)


# for name in elites:
#     tweepy.Cursor(API.user_timeline, elites[name])


#test_tweets = API.user_timeline(screen_name='IvanDuque', since_id=981697880766468097, count=20, tweet_mode='extended')

# print("Recent tweets:\n")
# for tweet in test_tweets:
#     print(tweet.full_text)
#     print()
