==> Code/4_saving_results.R <==
##################################
# 4_Saving Results -- PS 239T    #
# Author: Elisabeth Earley       #
# Date: 4/25/18                  #
##################################
setwd("C:/Users/bbits/Desktop/Documentss/1 SCHOOL/1 POLITICAL SCIENCE/final_project_ps239t/Results")

library(stringr)

#display and save graph
#to do: create a loop for this as well, to download reproducibly and elegantly.

facet_map
ggsave("facet.pdf")

full_map
ggsave("full_map.pdf")

velez_map
ggsave("velez.pdf")

fajardo_map
ggsave("fajardo.pdf")

petro_map
ggsave("petro.pdf")

duque_map
ggsave("duque.pdf")

lleras_map
ggsave("lleras.pdf")

santos_map
ggsave("santos.pdf")

delaCalle_map
ggsave("delaCalle.pdf")
